Refactor and remove requirement for initial model in root component
TreeRootComponent is now WSRootComponent. This now has an Option[R] for
the model so can start with None, and will show a supplied "noData"
ReactElement until data is received from the server.
Add SimpleRootComponent that just manages a model instance with no
websocket (e.g. for testing of display and editing of a model without
server connection).
Remove unused import.