Support zooming to an index in a list, start of support for general lenses
We can implement support for specific lenses (e.g. LensN, a combination
of a Lens and a name) in Cursor using specific zoomXXX functions. These
changes should support using any Lens with a circe Decoder and an
OuterEncoder. OuterEncoder just allows encoding of nested
lenses/actions from the bottom up, as done in Cursor. This could
replace the specific zoomXXX methods, but at the moment type inference
makes this awkward - can't infer the child type.

Json encoding of deltas is now more consistent - we encode lenses as
first "lens" then say "lensN".

zoomI, OptionalI, optionalIDelta and optionalIParent now allow for
zooming to an index in a list. Again the optionalIDelta and
optionalIParent are really only for convenience - an encoder/decoder
pair is all that's really needed, but by using specific implementations
we make type inference better for common cases.

Add demo of zoomI for todo lists - we can now edit todos by index in
the list. Would be better to edit by id - this will probably be next
zoomXXX function ;)