DeltaCodecs to allow for ref tracking/updating
Add Refs - a Guid and a long Revision (may be enhanced in future so
that the revision uses a DeltaId). They refer to a particular revision
of a data item with the corresponding Guid.
A Cache provides for looking up a Ref's data.
Instead of Decoder[Delta[A]] we now use DeltaCodec[A], which provides a
Delta decoder, but also allows for updating Refs in a data item using a
Cache.
Use our string format for coding/decoding guids to make them shorter
and neater.
Add spec for refs to try out updating them, and Cache behaviour.