Remove automatic casting to Ref type in Cache
Before, Cache.apply(ref) would return data cast to the parametric type
of the reference. However this made it difficult to safely serialise
deltas on the cache. Therefore Cache.apply now returns data as the
parametric type of the cache. However we can easily use prisms to get
back to the type of the reference, using new Cache methods getPrism and
prism (which work like get and apply, but use an implicit prism). This
needs to be extended to CacheCursor and demonstrated in
tree-material-ui.