Move Ref to own file. Make Cache lookup ignore revisions.
Ignore revisions in cache lookup by Ref - we now always provide the
most recent revision of data, and revisions are just used to cause a
data change in React. This is necessary since server and client may
have different revision numbers. A better solution might be to use
guids for revisions.