Make Cursor include CursorP
Renamed p to location, makes sense to have a location for all Cursors,
it can be Unit if not needed. This gets rid of a lot of code, and since
CursorP is used nearly everywhere it makes sense to make it the default.
CursorL is now not needed - just use a string location.
Might make sense to name the field "p" again rather than location,
since it is very occasionally used for some other function, will
revisit this if it's a problem.