Move from Cache to Mirror
This has no parametric type. Ref[A] will retrieve an A, and we handle
codecs by storing inside the mirror, allowing deltas to work. Requires
a map from MirrorType (just a string used as a type tag) to MirrorCodec
(codec and DeltaCodec for a) to decode. Can encode without this using
codecs stored in the mirror.