Add support for adding refs at the top level in a DeltaIO
This uses a new action in DeltaIO that contains a function from a guid
to a new data item, and registers that item under the guid when run.

A typeclass RefAdder is used to actually update models to contain refs,
and we provide an implicit implementation for Mirror and a non-implicit
def giving a no-op for any other data type.

This involves adding the RefAddr typeclass in a number of places, when
using the library this is just needed for making ServerStore and
ServerRootComponents. We also make DeltaIORun methods return a new
result case class that includes the direct result of the DeltaIO, but
also a list of added references.