Reuse the same MirrorAndId.mirror lens
MirrorAndId.mirror lens is provided by a def, so provides a new
instance per invocation. This is because it accepts the MirrorAndId
parametric type as part of the parent type of the lens. We can avoid
this by getting an instance and keeping it in MirrorAndIdRootSource.
Since this is pretty much the only use of the lens we can make sure we
don't trigger unecessary React updates when the lens is included in a
Cursor.