Introduce Action as callback indirection layer
See `Action` scaladoc for rationale.
Cursor now returns Actions for `act` and `set`, and new `action` method
is equivalent to `callback` but returning an `Action`. `callback` is
retained for compatibility and to implement `Parent`.
Tested with `RefViews` in tree-material-ui and allows reuse of list
item components when updating, even though they contain delete actions.