Update Logoot to pass scalacheck
There was a typo in Position.apply that made it produce only empty
lists :) A more subtle thing was that we need to use the minimum of
qIdent and firstPosition.head to pad p - we could have qIdent with pos
0 and a negative client id. Actually the current code might not
generate that, but it's nice to reduce the number of invariants
required on ident, and it might be nice to make future code assign
clientId to padding idents - this seems to be what the logoot paper
aims for.

This is probably fine now, it will run 100,000 random position pairs
that should be chosen from any position that can actually be generated
by positionsBetween, of length 64 or less (and there's no real reason
to think there are actually any new interesting cases after a few
positions in length). Note that positionsBetween (by inspection)
guarantees an invariant on generated positions that their last
identifier has pos > 1, and relies on this to work. We therefore
enforce this in the random test positions.

However "proving" it is fine is challenging. There's a lot of comments
trying to set this out, but with some gaps and not all linked together
:(

Committing now as a baseline that is probably correct.