Add type of Delta as D <: Delta[A] wherever we previously used Delta[A]
This allows us to encode deltas directly without having to provide
encoders for Delta[A] - we can use Circe-generated codecs on ADTs of
deltas for each data type, rather than trying to pull stuff out of a
soup of possibly overlapping type ids in the Json. This should also
remove the need to pass around the Json for deltas as we build them in
cursors.

Have done some of the refactoring for this, more to go.