Introduce DLens and DOptional, and ListIndexDelta
These look a little like Lens and Optional. They provide the same A =>
B function as a Lens, but in the other direction they transform from a
delta of B to a delta of A. This is what is required to transform
models and deltas when zooming a Cursor through a data model.

ListIndexDelta provides insertion, deletion, updating and editing of
elements at a given index of a List.

Reimplemented part of ListCursor as ListIndexCursor, which provides
additional methods on a Cursor to a List using ListIndexDelta. These
are convenience methods to zoom to an index, and zoom to all indices (a
little like sequence - this more or less goes from Cursor[List] to
List[Cursor].