Add encoder to DeltaCodec, so we can remove serialisation in Parent
This means we can serialise deltas wherever we want, in particular we
should be able to implement OT by allowing deltas to transform each
other, and the serialise the transformed deltas. This also makes it
easier to keep client and server in sync, in terms of supported deltas.
Drop LensN and just supply names when making DeltaCodecs based on
lenses instead. Add names to other DeltaCodecs - in actions, this
allows us to distinguish multiple types of actions if needed, and in
match deltas it allows us to distinguish types of "finder". Finally,
add delta based on finding by Id, which should be used in preference to
generic match delta in nearly all cases.